backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
aiosqlite==0.20.0

# ML / Computer Vision
insightface==0.7.3
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# Export
openpyxl==3.1.5

# Utilities
pydantic==2.9.0
python-dateutil==2.9.0
